Update sequalize object model to parse float value (#62)

+++ src_logic/Entity/Prescription.js
@@ -7,10 +7,10 @@ const Prescription = dbService.define('prescription', {
   chartNumber	: { type :	sequelize.INTEGER(12), allowNull : false },
   medicineName : { type : sequelize.STRING(40), allowNull : false },
   medicineIngredient : { type : sequelize.STRING(100), allowNull : false },
-  doses : { type : sequelize.INTEGER(5), allowNull : true },
+  doses : { type : sequelize.FLOAT, allowNull : true },
   dosesCountByDay : { type : sequelize.CHAR(3), allowNull : true },
   dosesDay : { type : sequelize.INTEGER(5), allowNull : true },
-  useTotal : { type : sequelize.INTEGER(5), allowNull : false },
+  useTotal : { type : sequelize.FLOAT, allowNull : false },
   useFlag : { type : sequelize.BOOLEAN, allowNull : false, defaultValue : 0 },
   remarks : { type : sequelize.STRING(100), allowNull : true }
 });
